eal:vfio reorganized spapr code for Power
Reorganized the sPAPR VFIO code to make it easier to read. Modified
the DMA window create/remove logic so that the window only grows, never
shrinks. This should reduce the number of unmap/unregister/remove/
create/register/map operations required. Added support for IOVA=TA mode
such that the DMA window is created for exactly the size requested on
the command line, not dynamically sized based on memory requests.

Signed-off-by: David Christensen <drc@linux.vnet.ibm.com>